refactor(report): promote transaction stats to a subdomain
Restructure the report ドメイン so the transaction-stats slice lives
as a sibling subdomain of `report/feedback/`, matching the pattern
already used across the codebase (e.g. `account/community/config/`,
`transaction/incentiveGrant/`, `experience/evaluation/evaluationHistory/`).

  src/application/domain/report/
    transactionStats/         # ★ new subdomain
      data/
        repository.ts         # 13 aggregation + MV-refresh methods
        interface.ts          # IReportTransactionStatsRepository (split out from the repo file)
        rows.ts               # 12 raw-SQL row types (moved from data/rows.ts)
      weeklyAggregator.ts     # 7 pure functions for the weekly payload

The existing data-layer split (`data/repository/transactionStats.ts`
under the parent) had no precedent in other domains. Moving it under
its own subdomain folder keeps the layered file pattern
(`<subdomain>/data/{repository, interface, rows}.ts` +
`<subdomain>/<helper>.ts`) aligned with how subdomains are organised
elsewhere. Implementation bodies are unchanged; this commit is purely
file relocation + import-path updates in 5 caller files
(service.ts / provider.ts / presenter/index.ts and 2 unit tests).
