+++ src/application/domain/report/transactionStats/data/interface.ts
@@ -1,0 +1,91 @@
+import { Prisma } from "@prisma/client";
+import { IContext } from "@/types/server";
+import {
+  CohortRetentionRow,
+  CommunityContextRow,
+  DateRange,
+  DeepestChainRow,
+  PeriodAggregateRow,
+  RetentionAggregateRow,
+  TransactionActiveUsersDailyRow,
+  TransactionCommentRow,
+  TransactionSummaryDailyRow,
+  UserProfileForReportRow,
+  UserTransactionAggregateRow,
+} from "@/application/domain/report/transactionStats/data/rows";
+
+/**
+ * Transaction-statistics repository contract for the report domain.
+ * Aggregations + materialized-view refreshes are scoped here so callers
+ * (ReportService.weeklyPayload, sysadmin retention/cohort orchestrators)
+ * can ask for a single concern without dragging the entity / template
+ * surface around.
+ */
+export interface IReportTransactionStatsRepository {
+  findDailySummaries(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+  ): Promise<TransactionSummaryDailyRow[]>;
+  findDailyActiveUsers(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+  ): Promise<TransactionActiveUsersDailyRow[]>;
+  findTopUsersByTotalPoints(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+    topN: number,
+  ): Promise<UserTransactionAggregateRow[]>;
+  findTrueUniqueCounterpartiesForUsers(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+    userIds: string[],
+  ): Promise<Map<string, number>>;
+  findCommentsByDateRange(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+    limit?: number,
+  ): Promise<TransactionCommentRow[]>;
+  findUserProfiles(
+    ctx: IContext,
+    communityId: string,
+    userIds: string[],
+  ): Promise<UserProfileForReportRow[]>;
+  findCommunityContext(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+  ): Promise<CommunityContextRow | null>;
+  findDeepestChain(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+  ): Promise<DeepestChainRow | null>;
+  findPeriodAggregate(
+    ctx: IContext,
+    communityId: string,
+    range: DateRange,
+  ): Promise<PeriodAggregateRow>;
+  findRetentionAggregate(
+    ctx: IContext,
+    communityId: string,
+    range: {
+      currentWeekStart: Date;
+      nextWeekStart: Date;
+      prevWeekStart: Date;
+      twelveWeeksAgo: Date;
+    },
+  ): Promise<RetentionAggregateRow>;
+  findCohortRetention(
+    ctx: IContext,
+    communityId: string,
+    cohort: { cohortStart: Date; cohortEnd: Date },
+    active: { activeStart: Date; activeEnd: Date },
+  ): Promise<CohortRetentionRow>;
+  refreshTransactionSummaryDaily(ctx: IContext, tx: Prisma.TransactionClient): Promise<void>;
+  refreshUserTransactionDaily(ctx: IContext, tx: Prisma.TransactionClient): Promise<void>;
+}
