+++ src/application/domain/report/transactionStats/data/rows.ts
